feat: connection dialog
